+++ packages/button/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
-## [0.4.1](https://github.com/Availity/element/compare/@availity/mui-button@0.4.0...@availity/mui-button@0.4.1) (2023-10-03)
+## [0.4.2](https://github.com/Availity/element/compare/@availity/mui-button@0.4.1...@availity/mui-button@0.4.2) (2023-10-06)
 
 ### Bug Fixes