+++ packages/checkbox/CHANGELOG.md
@@ -2,11 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.1...@availity/mui-checkbox@0.1.2) (2023-10-17)
+
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-checkbox@0.1.0...@availity/mui-checkbox@0.1.1) (2023-10-04)
 
 ## 0.1.0 (2023-10-04)
 
-
 ### Features
 
-* **mui-checkbox:** init ([c7200a8](https://github.com/Availity/element/commit/c7200a84175174f43d502b47881dd7e57e02aa0e))
+- **mui-checkbox:** init ([c7200a8](https://github.com/Availity/element/commit/c7200a84175174f43d502b47881dd7e57e02aa0e))