+++ packages/tabs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.3...@availity/mui-tabs@0.1.4) (2023-10-06)
+
 ## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.2...@availity/mui-tabs@0.1.3) (2023-08-23)
 
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-tabs@0.1.1...@availity/mui-tabs@0.1.2) (2023-08-22)