+++ packages/tooltip/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.5.3](https://github.com/Availity/element/compare/@availity/mui-tooltip@0.5.2...@availity/mui-tooltip@0.5.3) (2023-10-06)
+
 ## [0.5.2](https://github.com/Availity/element/compare/@availity/mui-tooltip@0.5.1...@availity/mui-tooltip@0.5.2) (2023-08-23)
 
 ## [0.5.1](https://github.com/Availity/element/compare/@availity/mui-tooltip@0.5.0...@availity/mui-tooltip@0.5.1) (2023-08-16)